requirements.txt
absl-py==2.1.0
accelerate==1.2.0
addict==2.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
array_record==0.5.1
asttokens==3.0.0
astunparse==1.6.3
attrs==24.2.0
bddl==3.5.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
cryptography==42.0.8
cycler==0.12.1
Cython==3.0.11
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
deepspeed==0.16.2
Deprecated==1.2.15
diffusers==0.11.1
dlimp @ git+https://github.com/moojink/dlimp_openvla@040105d256bd28866cc6620621a3d5f7b6b91b46
dm-tree==0.1.8
docker-pycreds==0.4.0
draccus==0.8.0
easydict==1.13
egl_probe==1.0.2
einops==0.8.0
etils==1.11.0
evdev==1.7.1
exceptiongroup==1.2.2
executing==2.1.0
fasteners==0.19
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.55.3
fsspec==2024.10.0
future==1.0.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
glfw==2.8.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
hjson==3.1.0
huggingface-hub==0.26.5
hydra-core==1.3.2
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipaddress==1.0.23
ipython==8.31.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json-numpy==2.1.0
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
jupytext==1.16.6
keras==2.15.0
kiwisolver==1.4.7
libclang==18.1.1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mergedeep==1.3.4
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.8
mujoco==3.2.6
mujoco-py==2.1.2.14
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.2
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
open3d==0.18.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
OpenEXR==3.3.2
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathtools==0.1.2
peft==0.11.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
promise==2.3
prompt_toolkit==3.0.48
protobuf==3.20.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
pynput==1.7.7
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.9
pyOpenSSL==24.3.0
pyparsing==3.2.0
pyquaternion==0.9.9
pytest==8.3.4
python-dateutil==2.9.0.post0
python-xlib==0.33
pytz==2024.2
PyYAML==6.0.2
pyyaml-include==1.4.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.4.5
schedule==1.2.2
scikit-learn==1.6.0
scipy==1.14.1
sentencepiece==0.1.99
sentry-sdk==2.19.2
setproctitle==1.3.4
shortuuid==1.0.13
six==1.17.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.15.0
tensorflow-addons==0.23.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.15.0
tensorflow-graphics==2021.12.3
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
termcolor==2.5.0
threadpoolctl==3.5.0
timm==0.9.10
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
trimesh==4.5.3
triton==2.2.0
typeguard==2.13.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.14.1
zipp==3.21.0
